requirements.txt
fastapi
uvicorn
crewai
chromadb
google-generativeai
python-dotenv
langchain-google-genai
langchain-community
